Fin Totale

+++ ChargerDonnees.h
@@ -1,6 +1,6 @@
 #ifndef CHARGERDONNEES_H_INCLUDED
 #define CHARGERDONNEES_H_INCLUDED
 
-int ChargerDonnees(item_t*** items, short* iNbItems, char* sNomFichier);
+int ChargerDonnees(item_t*** items, int* iNbItems, char* sNomFichier);
 
 #endif // CHARGERDONNEES_H_INCLUDED
